Add explicit height & width attributes for images
CSS properties work in conjunction with the height and width attributes in the HTML.
When attributes values aren't included, if an image takes longer to load than the rest of the page
contents, it won't take up any space on the page at first but will suddenly cause a drastic
shift of the other page contents once it does load in. Explicitly stating a height and width
prevents this from happening, as space will be “reserved” on the page and appear blank until
the image loads.